refactor(app-framework): move Placeholder into app-framework/ui
Hoists the React placeholder out of `composer-app/src/components/` into
`@dxos/app-framework/ui` so any host can use it. The component is now
brand-agnostic — it accepts a `logo` prop (either a `ReactNode` or a
render function `({ className }) => ReactNode` so SVG logos can apply the
animation classes directly). composer-app threads `<Composer />` from
`@dxos/brand` through a thin `ComposerPlaceholder` wrapper.

`@dxos/react-ui` and `@dxos/ui-theme` move from devDependencies to
dependencies in app-framework — the Placeholder needs `ThemeProvider` /
`defaultTx` / `mx`. The `window.__bootLoader` global type now lives next
to the component that calls `dismiss()`, so composer-app drops its
duplicate declaration.
