+++ packages/apps/composer-app/src/components/index.ts
@@ -2,5 +2,4 @@
 // Copyright 2023 DXOS.org
 //
 
-export * from './Placeholder';
 export * from './ResetDialog';
