+++ packages/sdk/app-framework/src/ui/components/Placeholder/index.ts
@@ -1,5 +1,5 @@
 //
-// Copyright 2023 DXOS.org
+// Copyright 2025 DXOS.org
 //
 
 export * from './Placeholder';
