+++ packages/sdk/app-framework/src/ui/components/index.ts
@@ -3,5 +3,6 @@
 //
 
 export * from './App';
+export * from './Placeholder';
 export * from './PluginManager';
 export * from './Surface';
